requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==23.12.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
Django==4.2
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pep8-naming==0.13.2
pillow==10.2.0
platformdirs==4.1.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
psycopg2-binary==2.9.9
psycopg==3.1.13
